move types to inline rather than module declaration
